requirements.txt
Django==4.2.7
celery==5.3.4
redis==5.0.1
django-ipware==3.0.7
requests==2.31.0
drf-yasg==1.21.5
django-ratelimit==3.0.1
gunicorn==21.2.0
django-cors-headers==4.3.1
requests==2.31.0
drf-yasg==1.21.5
django-ratelimit==3.0.1
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
dj-database-url==2.1.0
django-celery-results==2.5.1
django-celery-beat==2.5.0